+++ src/app/pages/home-page/offer-item-page/offer-item-page.component.spec.ts
@@ -2,7 +2,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 
 import { ApiService } from '@shared/services/api.service';
-import { MockApiService } from '@shared/services/mock-api.service';
+import { MockApiService } from '@shared/services/api.service.mock';
 
 import { OfferItemPageComponent } from './offer-item-page.component';
