+++ src/app/pages/home-page/offer-list-page/offer-list-page.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { ApiService } from '@shared/services/api.service';
-import { MockApiService } from '@shared/services/mock-api.service';
+import { MockApiService } from '@shared/services/api.service.mock';
 
 import { OfferListPageComponent } from './offer-list-page.component';
